removed: agenda component - fully integrated into the agenda_frame
