+++ src/frontend/__init__.py
@@ -1,9 +1,9 @@
-from .components import Agenda, Chat, ConnBox
+from .components import Chat, ConnBox
 from .agenda_frame import AgendaFrame
 from .chat_frame import ChatFrame
 from .layout import Layout
 from .modal import Modal
 
-__all__ = ["Agenda", "Chat", "ConnBox",
+__all__ = ["Chat", "ConnBox",
            "AgendaFrame", "ChatFrame", "Layout",
            "Modal"]
