+++ src/frontend/components/__init__.py
@@ -1,6 +1,5 @@
-from .agenda import Agenda
 from .chat import Chat
 from .conn_box import ConnBox
 from .interfaces import PresenceChangeable
 
-__all__ = ["Agenda", "Chat", "ConnBox", "PresenceChangeable"]
+__all__ = ["Chat", "ConnBox", "PresenceChangeable"]
